+++ tests/test_action.py
@@ -44,3 +44,7 @@ async def test_hmaction(
         parameter="LOCK_TARGET_LEVEL",
         value=1,
     )
+
+    call_count = len(mock_client.method_calls)
+    await action.send_value(1)
+    assert (call_count + 1) == len(mock_client.method_calls)