+++ changelog.md
@@ -1,6 +1,6 @@
 # Version 2023.2.5 (2023-02-05)
 - Fix GitHub build/publish
-- Update tests to ensure lock and siren send service call if parameter doesn't change
+- Update tests to ensure lock, siren, action and button send service call if parameter doesn't change
 
 # Version 2023.2.3 (2023-02-05)
 - Add comments to parameter_visibility